+++ fcn1.py
@@ -91,7 +91,7 @@ class FCN(TorchModelV2, nn.Module):
         state: List[TensorType],
         seq_lens: TensorType,
     ) -> (TensorType, List[TensorType]):
-      x=input_dict["obs"]
+      x=input_dict["obs"].type(torch.float32)
       x=x.type(torch.float32)
       out=self.f(x,state,seq_lens)
       return  out, state
